Some fixes for minor sloppiness, and a big fix: size of MIDI packets should not be stored in a byte, obviously!

+++ fb01emu_win32drv/src/winmm_drv.cpp
@@ -83,8 +83,8 @@ static CRITICAL_SECTION midiInLock; /* Critical section for MIDI In */
 typedef struct tagPendingData
 {
 	uint8_t* pData;
-	uint8_t length;
-	uint8_t sent;
+	size_t length;
+	size_t sent;
 	DWORD timeStamp;
 	tagPendingData* pNext;
 } PendingData;
